Renamed help.js -> get-involved.js and built the page out
